+++ js/cookies-regulation-services.js
@@ -1,6 +1,15 @@
 window.cookiesRegulationServices = {};
 
-//  Google Tag Manager
+function insertScript(url) {
+    const script = document.createElement('script');
+    script.async = true;
+    script.src   = url;
+
+    const head = document.getElementsByTagName('head')[0];
+    head.appendChild(script);
+}
+
+// Google Tag Manager
 window.cookiesRegulationServices.googleTagManager = {
     requiredOptions: ['id'],
     cookiesIdentifiers: ['^ga.*$', '^_ga.*$', '^_gc.*$', '^_gi.*$', '^_hj.*$', '^__utma.*$', '^__utmb.*$', '^__utmc.*$', '^__utmt.*$', '^__utmz.*$', '^__gads.*$'],
@@ -8,15 +17,54 @@ window.cookiesRegulationServices.googleTagManager = {
         window.dataLayer = window.dataLayer || [];
         window.dataLayer.push({
             'gtm.start': new Date().getTime(),
-            event: 'gtm.js'
+            'event':     'gtm.js'
         });
 
-        var scripts = document.getElementsByTagName('script')[0];
+        insertScript('https://www.googletagmanager.com/gtm.js?id=' + options.id);
+    }
+};
 
-        var googleTagManagerScript = document.createElement('script');
-        googleTagManagerScript.async = true;
-        googleTagManagerScript.src ='https://www.googletagmanager.com/gtm.js?id=' + options.id;
+// Google Analytics
+window.cookiesRegulationServices.googleAnalytics = {
+    requiredOptions: ['id'],
+    cookiesIdentifiers: ['^ga.*$', '^_ga.*$', '^_gc.*$', '^_gi.*$', '^_hj.*$', '^__utma.*$', '^__utmb.*$', '^__utmc.*$', '^__utmt.*$', '^__utmz.*$', '^__gads.*$'],
+    callback: function (options) {
+        window.dataLayer = window.dataLayer || [];
+        window.gtag = function gtag() {
+            window.dataLayer.push(arguments);
+        };
 
-        scripts.parentNode.insertBefore(googleTagManagerScript, scripts);
+        const additionalData = {anonymize_ip: false};
+
+        if (options.anonymize_ip === true) {
+            additionalData.anonymize_ip = true;
+        }
+
+        insertScript('https://www.googletagmanager.com/gtag/js?id=' + options.id);
+        gtag('js', new Date());
+        gtag('config', options.id, additionalData);
+    }
+};
+
+// Hotjar
+window.cookiesRegulationServices.hotjar = {
+    requiredOptions: ['id'],
+    cookiesIdentifiers: ['hjClosedSurveyInvites', '^_hj[a-zA-Z0-9]*$'],
+    callback: function (options) {
+        window._hjSettings = {
+            hjid: options.id,
+            hjsv: 6,
+        };
+
+        window.hj = window.hj || function() {
+            (window.hj.q = window.hj.q || []).push(arguments)
+        };
+
+        const url = 'https://static.hotjar.com/c/hotjar-'
+            + window._hjSettings.hjid
+            + '.js?sv='
+            + window._hjSettings.hjsv;
+
+        insertScript(url);
     }
 };